fix: [#1543] return always in API the downloads number from tracker-core
The tracker-core always has the metric alhoutght it can be persisted or
not. When it's not persisted, it contains the number of downloads during
the session.

On the other hand, the `torrent-repository` metri uses labels, so you
have to sum all values for all labels to get the total.

```
torrent_repository_torrents_downloads_total{peer_role="seeder"} 1
tracker_core_persistent_torrents_downloads_total{} 1
```
