+++ packages/axum-rest-tracker-api-server/src/v1/context/stats/handlers.rs
@@ -10,7 +10,6 @@ use bittorrent_udp_tracker_core::services::banning::BanService;
 use serde::Deserialize;
 use tokio::sync::RwLock;
 use torrust_rest_tracker_api_core::statistics::services::{get_labeled_metrics, get_metrics};
-use torrust_tracker_configuration::Core;
 
 use super::responses::{labeled_metrics_response, labeled_stats_response, metrics_response, stats_response};
 
@@ -41,10 +40,8 @@ pub struct QueryParams {
 #[allow(clippy::type_complexity)]
 pub async fn get_stats_handler(
     State(state): State<(
-        Arc<Core>,
         Arc<InMemoryTorrentRepository>,
         Arc<RwLock<BanService>>,
-        Arc<torrust_tracker_torrent_repository::statistics::repository::Repository>,
         Arc<bittorrent_tracker_core::statistics::repository::Repository>,
         Arc<bittorrent_http_tracker_core::statistics::repository::Repository>,
         Arc<torrust_udp_tracker_server::statistics::repository::Repository>,
@@ -57,8 +54,6 @@ pub async fn get_stats_handler(
         state.2.clone(),
         state.3.clone(),
         state.4.clone(),
-        state.5.clone(),
-        state.6.clone(),
     )
     .await;
