+++ packages/axum-rest-tracker-api-server/src/v1/context/stats/routes.rs
@@ -17,10 +17,8 @@ pub fn add(prefix: &str, router: Router, http_api_container: &Arc<TrackerHttpApi
         .route(
             &format!("{prefix}/stats"),
             get(get_stats_handler).with_state((
-                http_api_container.tracker_core_container.core_config.clone(),
                 http_api_container.tracker_core_container.in_memory_torrent_repository.clone(),
                 http_api_container.ban_service.clone(),
-                http_api_container.torrent_repository_container.stats_repository.clone(),
                 http_api_container.tracker_core_container.stats_repository.clone(),
                 http_api_container.http_stats_repository.clone(),
                 http_api_container.udp_server_stats_repository.clone(),
